requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cleanup==8.1.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
easy-thumbnails==2.8.5
Faker==25.8.0
idna==3.7
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
paypalrestsdk==1.13.3
pillow==10.3.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
